+++ crd-generator/api-v2/src/main/java/io/fabric8/crdv2/generator/AbstractJsonSchema.java
@@ -15,6 +15,7 @@
  */
 package io.fabric8.crdv2.generator;
 
+import com.fasterxml.jackson.annotation.JsonClassDescription;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.BeanDescription;
 import com.fasterxml.jackson.databind.BeanProperty;
@@ -370,7 +371,9 @@ public abstract class AbstractJsonSchema<T extends KubernetesJSONSchemaProps, V 
     }
 
     public void updateSchema(T schema) {
-      schema.setDescription(description);
+      if (Utils.isNotNullOrEmpty(description)) {
+        schema.setDescription(description);
+      }
       schema.setDefault(defaultValue);
       if (nullable) {
         schema.setNullable(true);
@@ -429,6 +432,11 @@ public abstract class AbstractJsonSchema<T extends KubernetesJSONSchemaProps, V 
 
     Class<?> rawClass = gos.javaType.getRawClass();
     collectDependentClasses(rawClass);
+
+    JsonClassDescription classDescription = rawClass.getAnnotation(JsonClassDescription.class);
+    if (classDescription != null) {
+      objectSchema.setDescription(classDescription.value());
+    }
 
     // while it should not be repeating, we reuse this method to look for preserve unknown on the class hierarchy
     consumeRepeatingAnnotation(rawClass, PreserveUnknownFields.class,